requirements.txt
# Web Framework
fastapi
uvicorn[standard]
pydantic
python-dotenv
jinja2

# Database
psycopg2-binary

# Auth & Security
passlib==1.7.4
bcrypt==3.2.2

# AI/ML Libraries
google-generativeai>=0.3.0
langchain
langchain-google-genai
langchain-openai
langchain-core
langchain-community

# JSON and Data Parsing
jsonschema